+++ src/components/active/edit-committee/committee-social-links/committee-social-links-list.tsx
@@ -2,8 +2,8 @@ import { useMemo, type FC } from "react";
 import { useFieldArray } from "react-hook-form";
 import type { z } from "zod";
 import { MAX_NUMER_OF_SOCIAL_LINKS } from "~/constants/committees";
+import { useAddCommitteeSocialLinksAsActive } from "~/hooks/mutations/useMutateCommittee";
 import { useFormWithZodSchema } from "~/hooks/useFormWithZodSchema";
-import { useAddCommitteeSocialLinksAsActive } from "~/hooks/useSetCommitteeSocialLinksAsActive";
 import { upsertCommitteeSocialLinksBaseSchema } from "~/server/api/helpers/schemas/committees";
 import CommitteeSocialLinksListActions from "./committee-social-links-list-actions";
 import CommitteeSocialLinksListItem from "./committee-social-links-list-item";
