+++ src/components/active/edit-committee/update-user-wizard.tsx
@@ -8,8 +8,8 @@ import { NumberInput } from "~/components/forms/NumberInput";
 import { Button } from "~/components/ui/button";
 import { Form } from "~/components/ui/form";
 import { MAX_ORDER_NUMBER, MIN_ORDER_NUMBER } from "~/constants/committees";
+import { useUpdateMemberAsActive } from "~/hooks/mutations/useMutateMember";
 import { useFormWithZodSchema } from "~/hooks/useFormWithZodSchema";
-import { useUpdateMemberAsActive } from "~/hooks/useUpdateMemberAsActive";
 import { upsertMemberBaseSchema } from "~/server/api/helpers/schemas/members";
 import type { RouterOutputs } from "~/utils/api";
 import localeObject from "~/utils/dayjs";
